Correct turrets
The game is an isometric view. When the car is standing upright, the tower moves up. The mount of the turret remains in place, and the turret itself leaves.
This commit corrects the bug.

It is impossible to change the departure of the turret from its position while moving. The graphics engine of the game does not allow you to achieve the desired effect.